third_party/WebKit/LayoutTests/http/tests/css/ignore-pending-link-import-expected.txt

backgroundColor: PASS
color: PASS
backgroundColor: PASS
color: PASS
